refactor: refactoring for typescript
