+++ src/components/Logo/Logo.styles.ts
