+++ src/pages/Specialties/Specialties.styles.ts
